Extract Tracer interface, java-tracing and reflective implementations
